feat(tasks): Update iOS tasks
